Angular_FastAPI_Version/Backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf2
python-docx
pdfplumber
requests
python-multipart